feat: move DragDropContext to FormBuilderPage
to be shared between builder drawer and content
