+++ frontend/src/features/admin-form-builder/BuilderContent.tsx
@@ -1,5 +1,5 @@
-import { memo, useCallback, useEffect } from 'react'
-import { DragDropContext, Droppable, DropResult } from 'react-beautiful-dnd'
+import { memo, useEffect } from 'react'
+import { Droppable } from 'react-beautiful-dnd'
 import { Flex, Stack } from '@chakra-ui/react'
 
 import { AdminFormDto } from '~shared/types/form'
@@ -8,38 +8,10 @@ import { useAdminForm } from '~features/admin-form/common/queries'
 
 import { FieldRowContainer } from './FieldRow/FieldRowContainer'
 import { clearActiveFieldSelector, useEditFieldStore } from './editFieldStore'
-import { useMutateFormFields } from './mutations'
 
 export const BuilderContent = (): JSX.Element => {
   const clearActiveField = useEditFieldStore(clearActiveFieldSelector)
   const { data } = useAdminForm()
-  const { mutateReorderField } = useMutateFormFields()
-
-  const onBeforeCapture = useCallback(() => {
-    /*...*/
-  }, [])
-  const onBeforeDragStart = useCallback(() => {
-    /*...*/
-  }, [])
-  const onDragStart = useCallback(() => {
-    /*...*/
-  }, [])
-  const onDragUpdate = useCallback(() => {
-    /*...*/
-  }, [])
-  const onDragEnd = useCallback(
-    ({ source, destination }: DropResult) => {
-      if (!data || !destination || destination.index === source.index) {
-        return
-      }
-      return mutateReorderField.mutate({
-        fields: data.form_fields,
-        from: source.index,
-        to: destination.index,
-      })
-    },
-    [data, mutateReorderField],
-  )
 
   useEffect(() => {
     // Clear field on component unmount.
@@ -65,26 +37,18 @@ export const BuilderContent = (): JSX.Element => {
           w="100%"
           flexDir="column"
         >
-          <DragDropContext
-            onBeforeCapture={onBeforeCapture}
-            onBeforeDragStart={onBeforeDragStart}
-            onDragStart={onDragStart}
-            onDragUpdate={onDragUpdate}
-            onDragEnd={onDragEnd}
-          >
-            <Droppable droppableId="formFieldList">
-              {(provided) => (
-                <Stack
-                  spacing="2.25rem"
-                  ref={provided.innerRef}
-                  {...provided.droppableProps}
-                >
-                  <BuilderFields fields={data?.form_fields} />
-                  {provided.placeholder}
-                </Stack>
-              )}
-            </Droppable>
-          </DragDropContext>
+          <Droppable droppableId="formFieldList">
+            {(provided) => (
+              <Stack
+                spacing="2.25rem"
+                ref={provided.innerRef}
+                {...provided.droppableProps}
+              >
+                <BuilderFields fields={data?.form_fields} />
+                {provided.placeholder}
+              </Stack>
+            )}
+          </Droppable>
         </Flex>
       </Flex>
     </Flex>
